requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv
gino
sqlalchemy